+++ .travis.yml
@@ -1,5 +1,4 @@
 language: node_js
 node_js:
-  - "8"
   - "10"
   - "12"
